\name{MSO-package}
\alias{MSO-package}
\alias{MSO}
\docType{package}

\title{MSO}

\description{Fits various Bayesian multi-species occupancy models}

\details{The following package allows the user to build various Bayesian
  multi-species occupancy models.}
\author{A.E. Clark (allan.clark@uct.ac.za)}

\references{
%update this later!
\itemize{
\item MacKenzie, D.I., Nichols, J.D., Lachman, G.B., Droege, S., Royle, J.A., Langtimm, C.A. (2002). Estimating site occupancy rates when detection probabilities are less than one. Ecology, 83(8): 2248-2255}
}

% Optionally other standard keywords, one per line, from file %
% KEYWORDS in the R documentation directory %

